Move default server implementation to its own file
